Code cleanup, refactoring, README Added
