+++ models/experimental/SSD512/README.md
@@ -1,0 +1,114 @@
+# SSD512
+
+### Platforms: Wormhole (n150)
+### Supported Input Resolution:** `(512, 512)` = (Height, Width)
+
+## Introduction
+SSD512 (Single Shot MultiBox Detector) is a real-time object detection model that performs object detection in a single forward pass. The model uses a VGG backbone with additional extra layers and multibox heads to detect objects at multiple scales. SSD512 is implemented for 512x512 input images and can detect 21 object classes (VOC dataset format).
+
+## Prerequisites
+- Cloned [tt-metal repository](https://github.com/tenstorrent/tt-metal) for source code
+- Installed: [TT-Metalium™ / TT-NN™](https://github.com/tenstorrent/tt-metal/blob/main/INSTALLING.md)
+
+## Setup
+
+### Download Weights (Optional)
+The demo currently uses random weights for testing. To use trained weights:
+1. Download trained SSD512 weights (VOC format)
+2. Load weights into the PyTorch model before transferring to TTNN
+
+
+## Repository Layout
+```
+models/
+└── experimental/
+    └── SSD512/
+        ├── resources/
+        │   ├── sample_input/
+        │   │   └── 000007.jpg
+        │   └── sample_output/
+        ├── reference/
+        │   ├── ssd.py                    # TorchSSD (reference)
+        │   ├── data/
+        │   │   ├── voc0712.py
+        │   │   ├── config.py
+        │   │   └── scripts/
+        │   └── layers/
+        │       ├── box_utils.py
+        │       ├── functions/
+        │       │   ├── detection.py
+        │       │   └── prior_box.py
+        │       └── modules/
+        │           ├── l2norm.py
+        │           └── multibox_loss.py
+        ├── tt/
+        │   ├── tt_ssd.py                 # SSD512Network (TTNN)
+        │   └── layers/
+        │       ├── detect.py
+        │       ├── tt_vgg_backbone.py
+        │       ├── tt_extras_backbone.py
+        │       ├── tt_l2norm.py
+        │       ├── tt_multibox_heads.py
+        │       └── functions/
+        │           └── prior_box.py
+        ├── demo/
+        │   └── demo.py                   # CLI demo
+        ├── tests/
+        │   ├── pcc/
+        │   │   ├── test_ssd.py           # end-to-end pytest
+        │   │   ├── test_vgg_backbone.py
+        │   │   ├── test_extras_backbone.py
+        │   │   ├── test_l2norm.py
+        │   │   └── test_multibox_heads.py
+        │   ├── perf/
+        │       └── test_device_perf_ssd.py
+        └── README.md
+```
+
+## Details
+
+- The entry point to the TTNN SSD512 model is `SSD512Network` in `models/experimental/SSD512/tt/tt_ssd.py`. The model uses random weights from the PyTorch reference implementation.
+
+## How to Run
+
+### Run the Full Model Test
+```bash
+# From tt-metal root directory
+pytest models/experimental/SSD512/tests/pcc/test_ssd.py
+```
+
+### Performance
+### Single Device (BS=1):
+- end-2-end perf is `47.1` FPS
+
+### Run Device Performance Test
+```bash
+# Test full model performance
+pytest models/experimental/SSD512/tests/perf/test_device_perf_ssd.py
+```
+
+### Run the Demo
+```bash
+# Process images from a directory
+python models/experimental/SSD512/demo/demo.py --input_dir <input_dir> --output_dir <output_dir>
+
+# With custom detection parameters
+python models/experimental/SSD512/demo/demo.py --input_dir <input_dir> --output_dir <output_dir> --conf_thresh 0.3 --max_detections 5
+```
+
+For help with demo options:
+```bash
+python models/experimental/SSD512/demo/demo.py --help
+```
+Note: Currently, the input image directory path for the demo: models/experimental/SSD512/resources/sample_input/
+
+### Demo Output Files
+
+The demo generates output files for each processed image:
+- `{image_name}_ttnn.jpg`: TTNN detection results with bounding boxes and labels
+Note: Default directory path for output image: models/experimental/SSD512/resources/sample_output/
+
+## Configuration Notes
+- Resolution: (H, W) = (512, 512) is supported end-to-end.
+- Device: The demo opens a Wormhole device (default id typically 0). If you need to change it, adjust the DemoConfig or the device open call in the demo.
+- Batch Size: Demo/tests are written for BS=1. For larger BS you’ll need to verify memory layouts and tile alignment.
